Update the copyright notices (drive-by), and introduce a couple of unit tests.
