+++ plugins/context_functions.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2016-2020 Jason Barrie Morley
+# Copyright (c) 2016-2021 Jason Barrie Morley
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
